+++ cmd/qbot/main.go
@@ -5,9 +5,14 @@ import (
 	"fmt"
 	"github.com/Insulince/jlib/pkg/jmain"
 	"github.com/Insulince/qbot/internal/qbot"
-	"log"
+	"github.com/pkg/errors"
+	"os"
 
 	_ "modernc.org/sqlite" // SQLite driver
+)
+
+const (
+	EnvVarDiscordBotToken = "DISCORD_BOT_TOKEN"
 )
 
 func main() {
@@ -15,17 +20,23 @@ func main() {
 }
 
 func Main(ctx context.Context) error {
-	// Initialize and start the bot
-	b, err := qbot.NewQBot()
-	if err != nil {
-		log.Fatalf("Error initializing bot: %v", err)
+	token, found := os.LookupEnv(EnvVarDiscordBotToken)
+	if !found {
+		return fmt.Errorf("environment variable %q not set", EnvVarDiscordBotToken)
 	}
 
+	// Initialize and start the bot
+	q, err := qbot.New(token)
+	if err != nil {
+		return errors.Wrap(err, "new qbot")
+	}
+
+	// Block main so the bot can run.
 	<-ctx.Done()
 
 	// Graceful shutdown
-	fmt.Println("Shutting down bot...")
-	b.Close()
+	fmt.Println("Q is shutting down...")
+	q.Close()
 
 	return nil
 }