+++ internal/qbot/handle-enter.go
@@ -6,7 +6,7 @@ import (
 	"time"
 )
 
-func (q *QBot) handleEnter(s *discordgo.Session, m *discordgo.MessageCreate) {
+func (q *QBot) handleEnter(m *discordgo.MessageCreate, _ []string) error {
 	q.queueMutex.Lock()
 	defer q.queueMutex.Unlock()
 
@@ -25,24 +25,18 @@ func (q *QBot) handleEnter(s *discordgo.Session, m *discordgo.MessageCreate) {
 
 		if inQueue {
 			// User is in queue but not active yet
-			s.ChannelMessageSend(m.ChannelID, fmt.Sprintf(
-				"<@%s>, you are currently in the queue at position %d. Please wait for your turn before using `!enter`.",
-				m.Author.ID, position))
+			q.mustPost(m.ChannelID, fmt.Sprintf("<@%s>, you are currently in the queue at position %d. Please wait for your turn before using `!enter`.", m.Author.ID, position))
 		} else {
 			// User is not in the queue at all
-			s.ChannelMessageSend(m.ChannelID, fmt.Sprintf(
-				"<@%s>, you are not currently in the queue. Use `!queue` to join.",
-				m.Author.ID))
+			q.mustPost(m.ChannelID, fmt.Sprintf("<@%s>, you are not currently in the queue. Use `!queue` to join.", m.Author.ID))
 		}
-		return
+		return nil
 	}
 
 	// If the user is already marked as entered, prevent redundant entry
 	if q.currentUser.Entered {
-		s.ChannelMessageSend(m.ChannelID, fmt.Sprintf(
-			"<@%s>, you have already entered your bracket. Please use `!full` when your bracket is complete.",
-			m.Author.ID))
-		return
+		q.mustPost(m.ChannelID, fmt.Sprintf("<@%s>, you have already entered your bracket. Please use `!full` when your bracket is complete.", m.Author.ID))
+		return nil
 	}
 
 	// Mark user as having entered
@@ -50,7 +44,7 @@ func (q *QBot) handleEnter(s *discordgo.Session, m *discordgo.MessageCreate) {
 	q.currentUser.AddedAt = time.Now()
 	q.currentUser.Warned = false
 
-	s.ChannelMessageSend(m.ChannelID, fmt.Sprintf(
-		"<@%s>, noted that you've entered your bracket. Please type `!full` once your bracket is full.",
-		m.Author.ID))
+	q.mustPost(m.ChannelID, fmt.Sprintf("<@%s>, noted that you've entered your bracket. Please type `!full` once your bracket is full.", m.Author.ID))
+
+	return nil
 }