+++ internal/qbot/handle-help.go
@@ -1,9 +1,12 @@
 package qbot
 
-import "github.com/bwmarrin/discordgo"
+import (
+	"github.com/bwmarrin/discordgo"
+	"github.com/pkg/errors"
+)
 
 // handleHelp displays a list of available commands.
-func (q *QBot) handleHelp(s *discordgo.Session, m *discordgo.MessageCreate) {
+func (q *QBot) handleHelp(m *discordgo.MessageCreate, args []string) error {
 	helpMessage := "" +
 		"**About**\n" +
 		"Q helps mediate a tourney queue to avoid joining the same bracket as other players.\n" +
@@ -21,7 +24,16 @@ func (q *QBot) handleHelp(s *discordgo.Session, m *discordgo.MessageCreate) {
 		"- If you take too long to `!enter` when you are at the front of the queue, you will be assumed AFK and kicked out of the queue and the next player will be given a turn. There is a 2 minute warning before this happens.\n" +
 		"- If you take too long to `!full` to signal your bracket is filled after doing `!enter`, you will be assumed AFK and kicked out of the queue and the next player will be given a turn. There is a 2 minute warning before this happens.\n" +
 		"- If you need more time because your bracket isn't full yet but you are close to being timed out, use `!moretime` to give yourself more time for this purpose.\n"
-	s.ChannelMessageSend(m.ChannelID, helpMessage)
-	q.handleCommands(s, m)
-	q.handleVersion(s, m)
+
+	q.mustPost(m.ChannelID, helpMessage)
+
+	if err := q.handleCommands(m, args); err != nil {
+		return errors.Wrap(err, "handle commands")
+	}
+
+	if err := q.handleVersion(m, args); err != nil {
+		return errors.Wrap(err, "handle version")
+	}
+
+	return nil
 }