+++ internal/qbot/handle-queue.go
@@ -7,20 +7,20 @@ import (
 )
 
 // handleQueue adds a user to the waiting queue.
-func (q *QBot) handleQueue(s *discordgo.Session, m *discordgo.MessageCreate) {
+func (q *QBot) handleQueue(m *discordgo.MessageCreate, _ []string) error {
 	q.queueMutex.Lock()
 	defer q.queueMutex.Unlock()
 
 	// Check if user is already active.
 	if q.currentUser != nil && q.currentUser.UserID == m.Author.ID {
-		s.ChannelMessageSend(m.ChannelID, fmt.Sprintf("<@%s>, you are already active. Please use `!enter` or `!full` as appropriate.", m.Author.ID))
-		return
+		q.mustPost(m.ChannelID, fmt.Sprintf("<@%s>, you are already active. Please use `!enter` or `!full` as appropriate.", m.Author.ID))
+		return nil
 	}
 	// Check if user is already in the waiting queue.
 	for _, item := range q.queue {
 		if item.UserID == m.Author.ID {
-			s.ChannelMessageSend(m.ChannelID, fmt.Sprintf("<@%s>, you are already in the queue.", m.Author.ID))
-			return
+			q.mustPost(m.ChannelID, fmt.Sprintf("<@%s>, you are already in the queue.", m.Author.ID))
+			return nil
 		}
 	}
 
@@ -33,7 +33,7 @@ func (q *QBot) handleQueue(s *discordgo.Session, m *discordgo.MessageCreate) {
 			Entered:   false,
 			Warned:    false,
 		}
-		s.ChannelMessageSend(m.ChannelID, fmt.Sprintf("<@%s>, it's your turn! Please type `!enter` once you join your bracket.", m.Author.ID))
+		q.mustPost(m.ChannelID, fmt.Sprintf("<@%s>, it's your turn! Please type `!enter` once you join your bracket.", m.Author.ID))
 	} else {
 		// Otherwise, add the user to the waiting queue.
 		q.queue = append(q.queue, QueueItem{
@@ -43,6 +43,8 @@ func (q *QBot) handleQueue(s *discordgo.Session, m *discordgo.MessageCreate) {
 			Entered:   false,
 			Warned:    false,
 		})
-		s.ChannelMessageSend(m.ChannelID, fmt.Sprintf("<@%s>, you've been added to the queue. Your position is %d.", m.Author.ID, len(q.queue)))
+		q.mustPost(m.ChannelID, fmt.Sprintf("<@%s>, you've been added to the queue. Your position is %d.", m.Author.ID, len(q.queue)))
 	}
+
+	return nil
 }