+++ internal/qbot/handle-reset.go
@@ -1,17 +1,21 @@
 package qbot
 
-import "github.com/bwmarrin/discordgo"
+import (
+	"github.com/bwmarrin/discordgo"
+)
 
 // handleReset clears the entire queue (moderator only).
-func (q *QBot) handleReset(s *discordgo.Session, m *discordgo.MessageCreate) {
-	if !isModerator(s, m) {
-		s.ChannelMessageSend(m.ChannelID, "You do not have permission to use this command. Moderator role required.")
-		return
+func (q *QBot) handleReset(m *discordgo.MessageCreate, _ []string) error {
+	if !q.isModerator(m) {
+		q.mustPost(m.ChannelID, "You do not have permission to use this command. Moderator role required.")
+		return nil
 	}
 
 	q.queueMutex.Lock()
 	defer q.queueMutex.Unlock()
 	q.currentUser = nil
 	q.queue = []QueueItem{}
-	s.ChannelMessageSend(m.ChannelID, "Queue has been reset.")
+	q.mustPost(m.ChannelID, "Queue has been reset.")
+
+	return nil
 }